LPS-94714 Properly export PortletURL, rename create_url to create_portlet_url and fix more JSDocs
