+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/index.es.js
@@ -16,21 +16,10 @@ import CompatibilityEventProxy from './liferay/CompatibilityEventProxy.es';
 import DefaultEventHandler from './liferay/DefaultEventHandler.es';
 import KeyboardFocusManager from './liferay/keyboard-focus/KeyboardFocusManager.es';
 import PortletBase from './liferay/PortletBase.es';
-import createURL from './liferay/util/portlet_url/create_url.es';
-import createActionURL from './liferay/util/portlet_url/create_action_url.es';
-import createRenderURL from './liferay/util/portlet_url/create_render_url.es';
-import createResourceURL from './liferay/util/portlet_url/create_resource_url.es';
 import fetch from './liferay/util/fetch.es';
 import navigate from './liferay/util/navigate.es';
 import ItemSelectorDialog from './liferay/ItemSelectorDialog.es';
 import objectToFormData from './liferay/util/form/object_to_form_data.es.js';
-
-const PortletURL = {
-	createActionURL,
-	createRenderURL,
-	createResourceURL,
-	createURL
-};
 
 export {AOP} from './liferay/aop/AOP.es';
 export {cancelDebounce, debounce} from './liferay/debounce/debounce.es';
@@ -43,6 +32,18 @@ export {PortletBase};
 export {Slider} from './liferay/compat/slider/Slider.es';
 export {Treeview} from './liferay/compat/treeview/Treeview.es';
 
+export {
+	default as createActionURL
+} from './liferay/util/portlet_url/create_action_url.es';
+export {
+	default as createPortletURL
+} from './liferay/util/portlet_url/create_portlet_url.es';
+export {
+	default as createRenderURL
+} from './liferay/util/portlet_url/create_render_url.es';
+export {
+	default as createResourceURL
+} from './liferay/util/portlet_url/create_resource_url.es';
 export {fetch};
 export {navigate};
 export {
@@ -50,4 +51,3 @@ export {
 } from './liferay/modal/commands/OpenSimpleInputModal.es';
 export {objectToFormData};
 export {openToast} from './liferay/toast/commands/OpenToast.es';
-export {PortletURL};
