+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/util/portlet_url/create_render_url.es.js
@@ -12,17 +12,17 @@
  * details.
  */
 
-import createURL from './create_url.es';
+import createPortletURL from './create_portlet_url.es';
 
 /**
- * Returns a render portlet URL in form of a string by setting the lifecycle parameter
+ * Returns a render portlet URL in form of a URL object by setting the lifecycle parameter
  * @param {!string} basePortletURL The base portlet URL to be modified in this utility
  * @param {object} parameters Search parameters to be added or changed in the base URL
- * @return {string} Render Portlet URL
+ * @return {URL} Render Portlet URL
  * @review
  */
 export default function createRenderURL(basePortletURL, parameters = {}) {
-	return createURL(basePortletURL, {
+	return createPortletURL(basePortletURL, {
 		...parameters,
 		p_p_lifecycle: '0'
 	});
