+++ modules/apps/frontend-js/frontend-js-web/test/liferay/util/portlet_url/create_portlet_url.es.js
@@ -14,9 +14,9 @@
 
 'use strict';
 
-import createURL from '../../../../src/main/resources/META-INF/resources/liferay/util/portlet_url/create_url.es';
+import createPortletURL from '../../../../src/main/resources/META-INF/resources/liferay/util/portlet_url/create_portlet_url.es';
 
-describe('Liferay.Util.PortletURL.createURL', () => {
+describe('Liferay.Util.PortletURL.createPortletURL', () => {
 	afterEach(() => {
 		Liferay.ThemeDisplay.getPortalURL.mockRestore();
 	});
@@ -30,14 +30,14 @@ describe('Liferay.Util.PortletURL.createURL', () => {
 	});
 
 	it('throws an error if basePortletURL is not a string', () => {
-		expect(() => createURL({portlet: 'url'}, {foo: 'bar'})).toThrow(
+		expect(() => createPortletURL({portlet: 'url'}, {foo: 'bar'})).toThrow(
 			'basePortletURL parameter must be a string'
 		);
 	});
 
 	it('throws an error if parameters is not an object', () => {
 		expect(() =>
-			createURL(
+			createPortletURL(
 				'http://localhost:8080/group/control_panel/manage',
 				'foo:bar'
 			)
@@ -46,14 +46,17 @@ describe('Liferay.Util.PortletURL.createURL', () => {
 
 	it('throws an error if parameters are provided but portlet ID is null', () => {
 		expect(() =>
-			createURL('http://localhost:8080/group/control_panel/manage', {
-				foo: 'bar'
-			})
+			createPortletURL(
+				'http://localhost:8080/group/control_panel/manage',
+				{
+					foo: 'bar'
+				}
+			)
 		).toThrow('Portlet ID must not be null if parameters are provided');
 	});
 
 	it('returns a URL object with a href parameter containing base url and parameters, where reserved parameters are not namespaced', () => {
-		const portletURL = createURL(
+		const portletURL = createPortletURL(
 			'http://localhost:8080/group/control_panel/manage?p_p_id=com_liferay_roles_admin_web_portlet_RolesAdminPortlet',
 			{
 				doAsGroupId: 'fooBar',
@@ -67,25 +70,25 @@ describe('Liferay.Util.PortletURL.createURL', () => {
 	});
 
 	it('returns a URL object when the base url is a mailto url', () => {
-		const portletURL = createURL('mailto:test@liferay.com');
+		const portletURL = createPortletURL('mailto:test@liferay.com');
 
 		expect(portletURL.href).toEqual('mailto:test@liferay.com');
 	});
 
 	it('returns a URL object consisting of a portal url followd by the base url when the base url  starts with a backslash', () => {
-		const portletURL = createURL('/foobar');
+		const portletURL = createPortletURL('/foobar');
 
 		expect(portletURL.href).toEqual('http://localhost:8080/foobar');
 	});
 
 	it('returns a URL object consisting of a portal url plus a slash followed by the base url when the base url is a basic string', () => {
-		const portletURL = createURL('foobar');
+		const portletURL = createPortletURL('foobar');
 
 		expect(portletURL.href).toEqual('http://localhost:8080/foobar');
 	});
 
 	it('overwrites a parameter in base url if the same parameter is set in the parameters argument', () => {
-		const portletURL = createURL(
+		const portletURL = createPortletURL(
 			'http://localhost:8080/group/control_panel/manage?p_p_id=com_liferay_roles_admin_web_portlet_RolesAdminPortlet&doAsGroupId=fooBar',
 			{
 				doAsGroupId: 'barBaz',
@@ -98,7 +101,7 @@ describe('Liferay.Util.PortletURL.createURL', () => {
 	});
 
 	it('prepends portal url if the provided base url does not have a proper beginning', () => {
-		const portletURL = createURL('/group/control_panel/manage');
+		const portletURL = createPortletURL('/group/control_panel/manage');
 
 		expect(portletURL.href).toEqual(
 			'http://localhost:8080/group/control_panel/manage'
